Refactor Jobs Controller
